Make needs_dynamic_casting multiple-complex-type aware.
We don't actually cast between these complex representations, but the prior implementation would indicate that we needed to dynamic_cast,
because we didn't have mappings for std::complex or thrust::complex.

This PR makes it so they all map to the same dtype.

Note that this has no functional change as all the use sites have already been changed to take this into account.

ghstack-source-id: 374b17d6d8b5b6d10a26ba7ba69fb297b3bf432c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39255