Add dynamic_cast asserts to CPU Loops.
On CUDA, we currently support casting loops dynamically (i.e. when the argument or return types of the lamba don't match the dtypes of the TensorIterator).
On CPU, before this change we would essentially reinterpret_cast, now we internal assert. We could add dynamic_casting support in the future on CPU.

ghstack-source-id: f747e90a5db620c6808472529bcdca7aa7719f7d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39258